+++ R/add.R
@@ -287,7 +287,7 @@ addMetaFeatures <- function(study, metaFeatures, overwrite = FALSE) {
 #' signature. The first argument is always \code{x}, which will be a data frame
 #' that combines the sample metadata with a column containing the assay
 #' measurements for a specific feature. That column will always be named
-#' \code{feature}. The second argument, \code{featureName}, is the unique ID of
+#' \code{feature}. The second argument, \code{featureID}, is the unique ID of
 #' the feature (idea: pass in the entire row from the features table so that the
 #' user can decide what metadata to display). This is for labeling the plot.
 #'